manual permission

+++ app/Http/Controllers/System/Settings/Pages/SocialLinkController.php
@@ -11,6 +11,8 @@ class SocialLinkController extends Controller
 {
     public function index(Request $request)
     {
+        $this->authorize('viewAny', SocialLink::class);
+
         $links = SocialLink::first();
         $phone_numbers = PhoneNumbers::select('id', 'phone_number')->get();
         return Inertia('System/Settings/Pages/SocialLinks/Index', [
@@ -21,6 +23,8 @@ class SocialLinkController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('viewAny', SocialLink::class);
+
         $phone_number = $request->validate([
             'phone_number' => 'required|unique:phone_numbers,phone_number',
             'user_id' => 'required|exists:users,id',
@@ -30,6 +34,8 @@ class SocialLinkController extends Controller
 
     public function update(Request $request)
     {
+        $this->authorize('viewAny', SocialLink::class);
+
         $data = $request->validate([
             'facebook'      => 'required|url|max:255',
             'instagram'       => 'required|url|max:1024',
@@ -46,6 +52,8 @@ class SocialLinkController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('viewAny', SocialLink::class);
+        
         $phoneNumber = PhoneNumbers::findOrFail($id);
 
         $phoneNumber->delete();
